requirements.txt
django==2.2.28
pyrebase4==4.5.0
